refactor: fixed support for public queries
