+++ src/api/get-positions-query.ts
@@ -2,7 +2,7 @@ import { gql, useQuery } from "@apollo/client";
 import { Position } from "cv-graphql";
 import { useNotificationContext } from "@/helpers/notification";
 
-const GET_POSITIONS_QUERY = gql`
+export const GET_POSITIONS_QUERY = gql`
     query Positions {
         positions {
             id
@@ -18,9 +18,11 @@ interface PositionsResult {
 export const usePositionsQuery = () => {
     const { showNotification } = useNotificationContext();
 
-    return useQuery<PositionsResult>(GET_POSITIONS_QUERY, {
+    const { data, ...query } = useQuery<PositionsResult>(GET_POSITIONS_QUERY, {
         onError: (error) => {
             showNotification("error", error.message);
         },
     });
+
+    return { positions: data?.positions ?? [], ...query };
 };
