+++ src/modules/departments/api/create-department-mutation.ts
@@ -1,7 +1,7 @@
 import { gql, useMutation } from "@apollo/client";
 import { CreateDepartmentInput } from "cv-graphql";
 import { CreateDepartmentResult } from "./departments.types";
-import { GET_DEPARTMENTS_QUERY } from "./get-departments-query";
+import { GET_DEPARTMENTS_QUERY } from "@/api";
 
 export const CREATE_DEPARTMENT = gql`
     mutation CreateDepartment($department: CreateDepartmentInput!) {
