+++ src/modules/departments/api/delete-department-mutation.ts
@@ -1,6 +1,6 @@
 import { gql, useMutation } from "@apollo/client";
 import { DeleteDepartmentInput } from "cv-graphql";
-import { GET_DEPARTMENTS_QUERY } from "./get-departments-query";
+import { GET_DEPARTMENTS_QUERY } from "@/api";
 
 export const DELETE_DEPARTMENT = gql`
     mutation DeleteDepartment($department: DeleteDepartmentInput!) {
