--- src/modules/departments/api/get-departments-query.ts
@@ -1,17 +1,0 @@
-import { gql, useQuery } from "@apollo/client";
-import { DepartmentsResult } from "./departments.types";
-
-export const GET_DEPARTMENTS_QUERY = gql`
-    query Departments {
-        departments {
-            id
-            name
-        }
-    }
-`;
-
-export const useDepartmentsQuery = () => {
-    const query = useQuery<DepartmentsResult>(GET_DEPARTMENTS_QUERY);
-
-    return { departments: query.data?.departments ?? [], ...query };
-};
