+++ src/modules/departments/api/index.ts
@@ -1,4 +1,3 @@
 export { useDepartmentCreate } from "./create-department-mutation";
 export { useDepartmentDelete } from "./delete-department-mutation";
 export { useDepartmentUpdate } from "./update-department-mutation";
-export { useDepartmentsQuery } from "./get-departments-query";
