+++ src/modules/departments/api/update-department-mutation.ts
@@ -1,7 +1,7 @@
 import { gql, useMutation } from "@apollo/client";
 import { UpdateDepartmentInput } from "cv-graphql";
 import { UpdateDepartmentResult } from "./departments.types";
-import { GET_DEPARTMENTS_QUERY } from "./get-departments-query";
+import { GET_DEPARTMENTS_QUERY } from "@/api";
 
 export const UPDATE_DEPARTMENT = gql`
     mutation UpdateDepartment($department: UpdateDepartmentInput!) {
